backend/compact-connect/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --no-emit-index-url compact-connect/requirements.in
#
attrs==23.2.0
    # via
    #   cattrs
    #   jsii
aws-cdk-asset-awscli-v1==2.2.202
    # via aws-cdk-lib
aws-cdk-asset-kubectl-v20==2.1.2
    # via aws-cdk-lib
aws-cdk-asset-node-proxy-agent-v6==2.0.3
    # via aws-cdk-lib
aws-cdk-aws-lambda-python-alpha==2.147.1a0
    # via -r compact-connect/requirements.in
aws-cdk-lib==2.147.1
    # via
    #   -r compact-connect/requirements.in
    #   aws-cdk-aws-lambda-python-alpha
    #   cdk-nag
cattrs==23.2.3
    # via jsii
cdk-nag==2.28.149
    # via -r compact-connect/requirements.in
constructs==10.3.0
    # via
    #   -r compact-connect/requirements.in
    #   aws-cdk-aws-lambda-python-alpha
    #   aws-cdk-lib
    #   cdk-nag
importlib-resources==6.4.0
    # via jsii
jsii==1.101.0
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v6
    #   aws-cdk-aws-lambda-python-alpha
    #   aws-cdk-lib
    #   cdk-nag
    #   constructs
publication==0.0.3
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v6
    #   aws-cdk-aws-lambda-python-alpha
    #   aws-cdk-lib
    #   cdk-nag
    #   constructs
    #   jsii
python-dateutil==2.9.0.post0
    # via jsii
six==1.16.0
    # via python-dateutil
typeguard==2.13.3
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v6
    #   aws-cdk-aws-lambda-python-alpha
    #   aws-cdk-lib
    #   cdk-nag
    #   constructs
    #   jsii
typing-extensions==4.12.2
    # via jsii
